+++ packages/ub/CHANGELOG.md
@@ -14,7 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+ - `RLS`: default `subjectIDsFn` function for the `aclRls` mixin didn't assign
+   org unit IDs of the current user into result
 
+## [5.23.4] - 2022-07-12
 ## [5.23.3] - 2022-07-11
 ### Added
  - `"Many" attribute storage entity`: added virtual ID attribute + mStorage mixin to be a subject of DDL generation